Book Seats in Restaurant
A website that allows customer to book seats in a restaurant of their choice. Designed this website to learn about integration concepts of front-end and back-end.
Copy the data-set provided in your MySQL Workbench and make the changes to the local host ports and addresses in ps3_connect.php accordingly.